fix for static build (ILLMMsgs)
