+++ src/devices/messages/ILLMMsgs/idl_generated_code/ILLMMsgs_index.txt
@@ -1,8 +1,8 @@
-yarp/dev/llm/return_readPrompt.h
-yarp/dev/llm/return_readPrompt.cpp
-yarp/dev/llm/return_ask.h
-yarp/dev/llm/return_ask.cpp
-yarp/dev/llm/return_getConversation.h
-yarp/dev/llm/return_getConversation.cpp
-yarp/dev/llm/ILLMMsgs.h
-yarp/dev/llm/ILLMMsgs.cpp
+yarp/dev/return_readPrompt.h
+yarp/dev/return_readPrompt.cpp
+yarp/dev/return_ask.h
+yarp/dev/return_ask.cpp
+yarp/dev/return_getConversation.h
+yarp/dev/return_getConversation.cpp
+yarp/dev/ILLMMsgs.h
+yarp/dev/ILLMMsgs.cpp
