+++ src/devices/messages/ILLMMsgs/idl_generated_code/yarp/dev/return_getConversation.h
@@ -8,29 +8,30 @@
 // This is an automatically generated file.
 // It could get re-generated if the ALLOW_IDL_GENERATION flag is on.
 
-#ifndef YARP_THRIFT_GENERATOR_STRUCT_RETURN_READPROMPT_H
-#define YARP_THRIFT_GENERATOR_STRUCT_RETURN_READPROMPT_H
+#ifndef YARP_THRIFT_GENERATOR_STRUCT_RETURN_GETCONVERSATION_H
+#define YARP_THRIFT_GENERATOR_STRUCT_RETURN_GETCONVERSATION_H
 
 #include <yarp/os/Wire.h>
 #include <yarp/os/idl/WireTypes.h>
+#include <yarp/dev/LLM_Message.h>
 #include <yarp/dev/ReturnValue.h>
 
-namespace yarp::dev::llm {
+namespace yarp::dev {
 
-class return_readPrompt :
+class return_getConversation :
         public yarp::os::idl::WirePortable
 {
 public:
     // Fields
     yarp::dev::ReturnValue ret{};
-    std::string prompt{};
+    std::vector<yarp::dev::LLM_Message> conversation{};
 
     // Default constructor
-    return_readPrompt() = default;
+    return_getConversation() = default;
 
     // Constructor with field values
-    return_readPrompt(const yarp::dev::ReturnValue& ret,
-                      const std::string& prompt);
+    return_getConversation(const yarp::dev::ReturnValue& ret,
+                           const std::vector<yarp::dev::LLM_Message>& conversation);
 
     // Read structure on a Wire
     bool read(yarp::os::idl::WireReader& reader) override;
@@ -48,7 +49,7 @@ public:
     std::string toString() const;
 
     // If you want to serialize this class without nesting, use this helper
-    typedef yarp::os::idl::Unwrapped<return_readPrompt> unwrapped;
+    typedef yarp::os::idl::Unwrapped<return_getConversation> unwrapped;
 
 private:
     // read/write ret field
@@ -57,13 +58,13 @@ private:
     bool nested_read_ret(yarp::os::idl::WireReader& reader);
     bool nested_write_ret(const yarp::os::idl::WireWriter& writer) const;
 
-    // read/write prompt field
-    bool read_prompt(yarp::os::idl::WireReader& reader);
-    bool write_prompt(const yarp::os::idl::WireWriter& writer) const;
-    bool nested_read_prompt(yarp::os::idl::WireReader& reader);
-    bool nested_write_prompt(const yarp::os::idl::WireWriter& writer) const;
+    // read/write conversation field
+    bool read_conversation(yarp::os::idl::WireReader& reader);
+    bool write_conversation(const yarp::os::idl::WireWriter& writer) const;
+    bool nested_read_conversation(yarp::os::idl::WireReader& reader);
+    bool nested_write_conversation(const yarp::os::idl::WireWriter& writer) const;
 };
 
-} // namespace yarp::dev::llm
+} // namespace yarp::dev
 
-#endif // YARP_THRIFT_GENERATOR_STRUCT_RETURN_READPROMPT_H
+#endif // YARP_THRIFT_GENERATOR_STRUCT_RETURN_GETCONVERSATION_H
