+++ src/devices/messages/ILLMMsgs/idl_generated_code/yarp/dev/return_readPrompt.h
@@ -8,30 +8,29 @@
 // This is an automatically generated file.
 // It could get re-generated if the ALLOW_IDL_GENERATION flag is on.
 
-#ifndef YARP_THRIFT_GENERATOR_STRUCT_RETURN_ASK_H
-#define YARP_THRIFT_GENERATOR_STRUCT_RETURN_ASK_H
+#ifndef YARP_THRIFT_GENERATOR_STRUCT_RETURN_READPROMPT_H
+#define YARP_THRIFT_GENERATOR_STRUCT_RETURN_READPROMPT_H
 
 #include <yarp/os/Wire.h>
 #include <yarp/os/idl/WireTypes.h>
-#include <yarp/dev/ILLM.h>
 #include <yarp/dev/ReturnValue.h>
 
-namespace yarp::dev::llm {
+namespace yarp::dev {
 
-class return_ask :
+class return_readPrompt :
         public yarp::os::idl::WirePortable
 {
 public:
     // Fields
     yarp::dev::ReturnValue ret{};
-    yarp::dev::LLM_Message answer{};
+    std::string prompt{};
 
     // Default constructor
-    return_ask() = default;
+    return_readPrompt() = default;
 
     // Constructor with field values
-    return_ask(const yarp::dev::ReturnValue& ret,
-               const yarp::dev::LLM_Message& answer);
+    return_readPrompt(const yarp::dev::ReturnValue& ret,
+                      const std::string& prompt);
 
     // Read structure on a Wire
     bool read(yarp::os::idl::WireReader& reader) override;
@@ -49,7 +48,7 @@ public:
     std::string toString() const;
 
     // If you want to serialize this class without nesting, use this helper
-    typedef yarp::os::idl::Unwrapped<return_ask> unwrapped;
+    typedef yarp::os::idl::Unwrapped<return_readPrompt> unwrapped;
 
 private:
     // read/write ret field
@@ -58,13 +57,13 @@ private:
     bool nested_read_ret(yarp::os::idl::WireReader& reader);
     bool nested_write_ret(const yarp::os::idl::WireWriter& writer) const;
 
-    // read/write answer field
-    bool read_answer(yarp::os::idl::WireReader& reader);
-    bool write_answer(const yarp::os::idl::WireWriter& writer) const;
-    bool nested_read_answer(yarp::os::idl::WireReader& reader);
-    bool nested_write_answer(const yarp::os::idl::WireWriter& writer) const;
+    // read/write prompt field
+    bool read_prompt(yarp::os::idl::WireReader& reader);
+    bool write_prompt(const yarp::os::idl::WireWriter& writer) const;
+    bool nested_read_prompt(yarp::os::idl::WireReader& reader);
+    bool nested_write_prompt(const yarp::os::idl::WireWriter& writer) const;
 };
 
-} // namespace yarp::dev::llm
+} // namespace yarp::dev
 
-#endif // YARP_THRIFT_GENERATOR_STRUCT_RETURN_ASK_H
+#endif // YARP_THRIFT_GENERATOR_STRUCT_RETURN_READPROMPT_H
